app/src/main/java/com/hasanalic/ecommerce/feature_home/domain/use_case/user_use_cases/DeleteUserUseCase.kt
package com.hasanalic.ecommerce.feature_home.domain.use_case.user_use_cases

import com.hasanalic.ecommerce.core.domain.model.DataError
import com.hasanalic.ecommerce.core.domain.model.Result
import com.hasanalic.ecommerce.feature_home.domain.repository.UserRepository
import javax.inject.Inject

class DeleteUserUseCase @Inject constructor(
    private val userRepository: UserRepository
) {
    suspend operator fun invoke(userId: String): Result<Unit, DataError.Local> {
        return userRepository.deleteUser(userId)
    }
}